Fix TestUtils_time with 32-bit time_t
tv_sec must be converted to uint64_t before the multiplication.

+++ runtime/core/Clownfish/TestHarness/TestUtils.c
@@ -148,7 +148,7 @@ TestUtils_time() {
     struct timeval t;
     gettimeofday(&t, NULL);
 
-    return (uint64_t)(t.tv_sec * 1000000 + t.tv_usec);
+    return (uint64_t)t.tv_sec * 1000000 + (uint64_t)t.tv_usec;
 }
 
 #else
